refactor: extract backend rate limiter mocks to own file
